feat: Add support for filtering tasks in getTasks endpoint
This commit modifies the `getTasks` endpoint in the `TasksController` to support filtering tasks based on status and search criteria. The endpoint now accepts a `filterDto` object as a query parameter, which can contain `status` and `search` properties. If the `filterDto` is provided, the `getTasksWithFilters` method of the `TasksService` is called to retrieve the filtered tasks. Otherwise, the `getAllTasks` method is used to retrieve all tasks.